Chore: painter가 들어간 아티스트 Composable 패키지 변경
